Fix Nagios core config creation in some cases
When the automatically created nagios.cfg in tmpfs is not available,
because the core was just switched or the tmpfs is not mounted, the
Nagios core config creation failed since a recent change, because
do_check_nagiosconfig() is now called in more situations, were it
simply can not succeed. Skip calling it in case the nagios.cfg is
not available

+++ cmk/base/core_config.py
@@ -293,7 +293,8 @@ def _backup_objects_file(core: MonitoringCore) -> Iterator[None]:
                 os.rename(backup_path, objects_file)
             raise
 
-        if config.monitoring_core == "nagios" and not do_check_nagiosconfig():
+        if (config.monitoring_core == "nagios" and
+                Path(cmk.utils.paths.nagios_config_file).exists() and not do_check_nagiosconfig()):
             broken_config_path = Path(cmk.utils.paths.tmp_dir) / "check_mk_objects.cfg.broken"
             shutil.move(cmk.utils.paths.nagios_objects_file, broken_config_path)
